backend/requirements.txt
flask
flask-cors
python-dotenv
pyjwt
flask-sqlalchemy
flask-migrate     
pymongo
flask-pymongo
werkzeug